# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pins:
    get:
      tags: [Pins]
      operationId: listPins
      summary: List pins visible to the current user
      description: >-
        Every role sees all pins. Field reps see other reps' pins as
        read-only context (rendered grey client-side). Managers/admins may
        pass `userId` to filter down to a single rep's pins.
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
          description: Managers/admins only — filter to a single rep's pins.
      responses:
        "200":
          description: Pins list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinListEnvelope"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      tags: [Pins]
      operationId: createPin
      summary: Drop a new pin
      description: Server reverse-geocodes the coordinates into a display address.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePinInput"
      responses:
        "201":
          description: Pin created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinEnvelope"
        "400":
          description: Invalid pin payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /pins/bulk:
    post:
      tags: [Pins]
      operationId: bulkCreatePins
      summary: Create multiple Roof pins from drone photos with embedded GPS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCreatePinsInput"
      responses:
        "201":
          description: Pins created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinListEnvelope"
        "400":
          description: Invalid payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /pins/{pinId}:
    delete:
      tags: [Pins]
      operationId: deletePin
      summary: Delete a pin
      description: Owners may delete their own pins; managers/admins may delete any pin.
      parameters:
        - name: pinId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pin deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteSuccess"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Not permitted to delete this pin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Pin not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /location/team:
    get:
      tags: [Location]
      operationId: listTeamLocations
      summary: List the latest known location of every team member
      description: Manager/admin only.
      responses:
        "200":
          description: Team locations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamLocationListEnvelope"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Manager/admin role required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    PinWorkflow:
      type: string
      enum: [retail, insurance]
    DamageType:
      type: string
      enum: [roof, siding, roof_and_siding]
    DoorKnockResult:
      type: string
      enum: [no_answer, no_appointment, appointment]
    ContactOutcome:
      type: string
      enum: [no_soliciting, priority_inspection, call_to_schedule]
...
    RetailData:
      type: object
      required:
        [
          ownerName1,
          interestedRoof,
          interestedSiding,
          interestedWindows,
          interestedDoors,
        ]
      properties:
        ownerName1:
          type: string
          minLength: 1
        ownerName2:
          type: [string, "null"]
        phone:
          type: [string, "null"]
        email:
          type: [string, "null"]
        interestedRoof:
          type: boolean
        interestedSiding:
          type: boolean
        interestedWindows:
          type: boolean
        interestedDoors:
          type: boolean
        interestNotes:
          type: [string, "null"]
        appointmentDate:
          type: [string, "null"]
        notes:
          type: [string, "null"]
    Pin:
      type: object
      required:
        [
          id,
          userId,
          latitude,
          longitude,
          address,
          workflow,
          damageType,
          photoUrl,
          doorKnockResult,
          retailData,
          contactOutcome,
          customerName,
          customerPhone,
          status,
          createdAt,
        ]
      properties:
        id:
          type: string
        userId:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        address:
          type: [string, "null"]
        workflow:
          $ref: "#/components/schemas/PinWorkflow"
        damageType:
          oneOf:
            - $ref: "#/components/schemas/DamageType"
            - type: "null"
        photoUrl:
          type: [string, "null"]
        doorKnockResult:
          oneOf:
            - $ref: "#/components/schemas/DoorKnockResult"
            - type: "null"
        retailData:
          oneOf:
            - $ref: "#/components/schemas/RetailData"
            - type: "null"
        contactOutcome:
          oneOf:
            - $ref: "#/components/schemas/ContactOutcome"
            - type: "null"
        customerName:
          type: [string, "null"]
        customerPhone:
          type: [string, "null"]
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    PinEnvelope:
      type: object
      required: [pin]
      properties:
        pin:
          $ref: "#/components/schemas/Pin"
    PinListEnvelope:
      type: object
      required: [pins]
      properties:
        pins:
          type: array
          items:
            $ref: "#/components/schemas/Pin"
    CreatePinInput:
      type: object
      required: [latitude, longitude, workflow]
      properties:
        latitude:
          type: number
        longitude:
          type: number
        workflow:
          $ref: "#/components/schemas/PinWorkflow"
        damageType:
          $ref: "#/components/schemas/DamageType"
        photoUrl:
          type: string
        doorKnockResult:
          $ref: "#/components/schemas/DoorKnockResult"
        retailData:
          $ref: "#/components/schemas/RetailData"
        contactOutcome:
          $ref: "#/components/schemas/ContactOutcome"
        customerName:
          type: string
        customerPhone:
          type: string
    BulkPinInput:
      type: object
      required: [latitude, longitude, photoUrl]
      properties:
        latitude:
          type: number
        longitude:
          type: number
        photoUrl:
          type: string
          minLength: 1
    BulkCreatePinsInput:
      type: object
      required: [pins]
      properties:
        pins:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/BulkPinInput"
    DeleteSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
...
    TeamLocation:
      type: object
      required: [userId, firstName, lastName, latitude, longitude, updatedAt]
      properties:
        userId:
          type: string
        firstName:
          type: [string, "null"]
        lastName:
          type: [string, "null"]
        latitude:
          type: number
        longitude:
          type: number
        updatedAt:
          type: string
          format: date-time
    TeamLocationListEnvelope:
      type: object
      required: [locations]
      properties:
        locations:
          type: array
          items:
            $ref: "#/components/schemas/TeamLocation"
...